catalog regression fix

+++ crates/catalog/src/store/flow/create.rs
@@ -122,7 +122,7 @@ impl CatalogStore {
 					constraint: column_to_create.constraint.clone(),
 					if_not_exists: false,
 					policies: vec![],
-					index: ColumnIndex(idx as u16),
+					index: ColumnIndex(idx as u8),
 					auto_increment: false,
 				},
 			)?;
